Bigint in most.

+++ db/migrate/20190611094224_create_pregnant_statuses.rb
@@ -3,7 +3,7 @@ class CreatePregnantStatuses < ActiveRecord::Migration[5.2]
     create_table :pregnant_statuses, :primary_key => :pregnant_status_id do |t|
      	t.bigint    :concept_id
       t.bigint      :encounter_id
-    	t.integer    :value_coded
+    	t.bigint    :value_coded
       t.boolean    :voided, null: false, default: 0
       t.bigint        :voided_by
       t.datetime   :voided_date
